Donbot/donbot_wordpress/requirements.txt
pika==1.1.0
python-wordpress-xmlrpc==2.3 
